+++ packages/nimble-components/src/rich-text/viewer/tests/rich-text-viewer.spec.ts
@@ -31,8 +31,8 @@ async function setupMentionConfig(): Promise<Fixture<RichTextViewer>> {
 
 async function appendUserMentionConfiguration(
     element: RichTextViewer,
-    userKeys: string[] | undefined,
-    displayNames: string[] | undefined
+    userKeys?: string[],
+    displayNames?: string[]
 ): Promise<void> {
     const userMention = document.createElement(
         richTextMentionUsersTag
@@ -128,8 +128,6 @@ describe('RichTextViewer', () => {
 
     describe('user mention dynamic loading', () => {
         beforeEach(async () => {
-            ({ element, connect, disconnect } = await setup());
-            pageObject = new RichTextViewerPageObject(element);
             await connect();
         });
 
@@ -147,7 +145,7 @@ describe('RichTextViewer', () => {
             expect(pageObject.getRenderedMarkdownLastChildContents()).toBe(
                 'user:1'
             );
-            await appendUserMentionConfiguration(element, undefined, undefined);
+            await appendUserMentionConfiguration(element);
 
             expect(pageObject.getRenderedMarkdownTagNames()).toEqual([
                 'P',
@@ -163,8 +161,8 @@ describe('RichTextViewer', () => {
         // TODO: Once the rich text validator added for duplicate configuration elements, below test case should be updated
         it('adding two mention configuration elements in the same viewer should render as mention view', async () => {
             element.markdown = '<user:1>';
-            await appendUserMentionConfiguration(element, undefined, undefined);
-            await appendUserMentionConfiguration(element, undefined, undefined);
+            await appendUserMentionConfiguration(element);
+            await appendUserMentionConfiguration(element);
 
             expect(pageObject.getRenderedMarkdownTagNames()).toEqual([
                 'P',
@@ -334,6 +332,38 @@ describe('RichTextViewer', () => {
                 )
             ).toEqual('username');
         });
+
+        it('updating invalid `key` in mapping mention should update it to absolute link if it is a mention before', async () => {
+            element.markdown = '<user:2>';
+            await appendUserMentionConfiguration(
+                element,
+                ['user:2'],
+                ['username']
+            );
+
+            expect(pageObject.getRenderedMarkdownTagNames()).toEqual([
+                'P',
+                `${richTextMentionUsersViewTag}`.toUpperCase()
+            ]);
+            expect(
+                pageObject.getRenderedMarkdownLastChildAttribute(
+                    'mention-label'
+                )
+            ).toEqual('username');
+
+            const renderedUserMention = element.firstElementChild as RichTextMentionUsers;
+            const renderedMappingUser = renderedUserMention.firstElementChild as MappingUser;
+            renderedMappingUser.key = 'invalid';
+            await waitForUpdatesAsync();
+
+            expect(pageObject.getRenderedMarkdownTagNames()).toEqual([
+                'P',
+                `${anchorTag}`.toUpperCase()
+            ]);
+            expect(pageObject.getRenderedMarkdownLastChildContents()).toBe(
+                'user:2'
+            );
+        });
     });
 
     describe('user mention via template', () => {
@@ -391,9 +421,17 @@ describe('RichTextViewer', () => {
     });
 
     describe('getMentionedHref() for viewer mentions', () => {
-        it('getMentionedHref() method should return the mentioned href when it valid mention configuration matching the pattern to mention node', async () => {
+        beforeEach(async () => {
+            await connect();
+        });
+
+        afterEach(async () => {
+            await disconnect();
+        });
+
+        it('should return the mentioned href when it valid mention configuration matching the pattern to mention node', async () => {
             element.markdown = '<user:1>';
-            await appendUserMentionConfiguration(element, undefined, undefined);
+            await appendUserMentionConfiguration(element);
             const renderedUserMention = element.firstElementChild as RichTextMentionUsers;
             expect(renderedUserMention.getMentionedHrefs()).toEqual(['user:1']);
         });
@@ -401,15 +439,15 @@ describe('RichTextViewer', () => {
         // TODO: Once the rich text validator added for duplicate configuration elements, below test case should be updated
         it('should return the mentioned href for duplicate mention configuration elements', async () => {
             element.markdown = '<user:1>';
-            await appendUserMentionConfiguration(element, undefined, undefined);
-            await appendUserMentionConfiguration(element, undefined, undefined);
+            await appendUserMentionConfiguration(element);
+            await appendUserMentionConfiguration(element);
             const renderedUserMention = element.firstElementChild as RichTextMentionUsers;
             expect(renderedUserMention.getMentionedHrefs()).toEqual(['user:1']);
         });
 
         it('should return unique mentioned href if same users exist twice', async () => {
             element.markdown = '<user:1> <user:1>';
-            await appendUserMentionConfiguration(element, undefined, undefined);
+            await appendUserMentionConfiguration(element);
             const renderedUserMention = element.firstElementChild as RichTextMentionUsers;
             expect(renderedUserMention.getMentionedHrefs()).toEqual([
                 'user:1'
@@ -418,7 +456,7 @@ describe('RichTextViewer', () => {
 
         it('should return all the mentioned href', async () => {
             element.markdown = '<user:1> <user:2>';
-            await appendUserMentionConfiguration(element, undefined, undefined);
+            await appendUserMentionConfiguration(element);
             const renderedUserMention = element.firstElementChild as RichTextMentionUsers;
             expect(renderedUserMention.getMentionedHrefs()).toEqual([
                 'user:1',